+++ src/main/java/HighThroughPutExchange/API/api_objects/requests/SetPriceRequest.java
@@ -6,12 +6,12 @@ import java.util.Map;
 
 public class SetPriceRequest extends BaseAdminRequest {
     @NotNull
-    private Map<String, Double> prices;
-    public SetPriceRequest(String adminUsername, String adminPassword, Map<String, Double> prices) {
+    private Map<String, Integer> prices;
+    public SetPriceRequest(String adminUsername, String adminPassword, Map<String, Integer> prices) {
         super(adminUsername, adminPassword);
         this.prices = prices;
     }
-    public Map<String, Double> getPrices() {
+    public Map<String, Integer> getPrices() {
         return prices;
     }
 }
