+++ src/main/java/HighThroughPutExchange/Common/OHLCData.java
@@ -1,8 +1,8 @@
 package HighThroughPutExchange.Common;
 
-public record OHLCData(double open, double high, double low, double close) {
+public record OHLCData(int open, int high, int low, int close) {
     @Override
     public String toString() {
-        return String.format("{Open: %.2f, High: %.2f, Low: %.2f, Close: %.2f}", open, high, low, close);
+        return String.format("{Open: %d, High: %d, Low: %d, Close: %d}", open, high, low, close);
     }
 }
