+++ src/main/java/HighThroughPutExchange/MatchingEngine/Order.java
@@ -3,11 +3,11 @@ package HighThroughPutExchange.MatchingEngine;
 public class Order {
     protected String name;
     protected String ticker;
-    protected double price;
-    protected double volume;
+    protected int price;
+    protected int volume;
     protected Side side;
     protected Status status;
-    public Order(String username, String ticker, double price, double volume, Side side, Status status) {
+    public Order(String username, String ticker, int price, int volume, Side side, Status status) {
         this.name = username;
         this.ticker = ticker;
         this.price = price;
@@ -16,7 +16,7 @@ public class Order {
         this.status = status;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
@@ -30,7 +30,7 @@ public class Order {
     @Override
     public String toString() {
         return String.format(
-                "Order{name='%s', ticker='%s', price=%.2f, volume=%.2f, side=%s, status=%s}",
+                "Order{name='%s', ticker='%s', price=%d, volume=%d, side=%s, status=%s}",
                 name, ticker, price, volume, side, status
         );
     }
