+++ src/main/java/HighThroughPutExchange/MatchingEngine/PriceChange.java
@@ -1,12 +1,12 @@
 package HighThroughPutExchange.MatchingEngine;
 
 public class PriceChange extends TradeKey{
-    private final double volume;
-    public PriceChange(String ticker, double price, double volume, Side side) {
+    private final int volume;
+    public PriceChange(String ticker, int price, int volume, Side side) {
         super(ticker, price, side);
         this.volume = volume;
     }
-    public double getVolume() {
+    public int getVolume() {
         return volume;
     }
     public int compareTo(PriceChange other) {
